+++ AKZ.API/DTOs/ProcessingDtos.cs
@@ -8,27 +8,27 @@ public class ProcessingRecordDto
     public string ProductMarker { get; set; } = string.Empty;
     public string WarehouseName { get; set; } = string.Empty;
     public string WorkerNames { get; set; } = string.Empty;
-    public int Count { get; set; }
-    public int RemainingCount { get; set; }
+    public double Count { get; set; }
+    public double RemainingCount { get; set; }
     public decimal UnitWeight { get; set; }
     public decimal RedWeight { get; set; }
-    public int RedCount { get; set; }
+    public double RedCount { get; set; }
     public decimal WhiteWeight { get; set; }
-    public int WhiteCount { get; set; }
+    public double WhiteCount { get; set; }
     public decimal SpecialWeight { get; set; }
-    public int SpecialCount { get; set; }
+    public double SpecialCount { get; set; }
     public decimal NaturalWeight { get; set; }
-    public int NaturalCount { get; set; }
+    public double NaturalCount { get; set; }
     public decimal NaturalWhiteWeight { get; set; }
-    public int NaturalWhiteCount { get; set; }
+    public double NaturalWhiteCount { get; set; }
     public decimal NaturalRedWeight { get; set; }
-    public int NaturalRedCount { get; set; }
+    public double NaturalRedCount { get; set; }
     public decimal ShortCutWeight { get; set; }
-    public int ShortCutCount { get; set; }
+    public double ShortCutCount { get; set; }
     public decimal ArtificialWeight { get; set; }
-    public int ArtificialCount { get; set; }
+    public double ArtificialCount { get; set; }
     public decimal ShortWeight { get; set; }
-    public int ShortCount { get; set; }
+    public double ShortCount { get; set; }
     public decimal LossWeight { get; set; }
     public decimal TotalWeight { get; set; }
     public decimal RemainingWeight { get; set; }
@@ -36,15 +36,15 @@ public class ProcessingRecordDto
     public decimal Difference { get; set; }
 
     // Remaining counts per category
-    public int RemRedCount { get; set; }
-    public int RemWhiteCount { get; set; }
-    public int RemSpecialCount { get; set; }
-    public int RemNaturalCount { get; set; }
-    public int RemNaturalWhiteCount { get; set; }
-    public int RemNaturalRedCount { get; set; }
-    public int RemShortCutCount { get; set; }
-    public int RemArtificialCount { get; set; }
-    public int RemShortCount { get; set; }
+    public double RemRedCount { get; set; }
+    public double RemWhiteCount { get; set; }
+    public double RemSpecialCount { get; set; }
+    public double RemNaturalCount { get; set; }
+    public double RemNaturalWhiteCount { get; set; }
+    public double RemNaturalRedCount { get; set; }
+    public double RemShortCutCount { get; set; }
+    public double RemArtificialCount { get; set; }
+    public double RemShortCount { get; set; }
 
     // Remaining weights per category
     public decimal RemRedWeight { get; set; }
@@ -64,27 +64,27 @@ public class CreateProcessingRecordDto
     public DateTime Date { get; set; }
     public int ProductId { get; set; }
     public string WorkerNames { get; set; } = string.Empty;
-    public int Count { get; set; }
-    public int RemainingCount { get; set; }
+    public double Count { get; set; }
+    public double RemainingCount { get; set; }
     public decimal UnitWeight { get; set; }
     public decimal RedWeight { get; set; }
-    public int RedCount { get; set; }
+    public double RedCount { get; set; }
     public decimal WhiteWeight { get; set; }
-    public int WhiteCount { get; set; }
+    public double WhiteCount { get; set; }
     public decimal SpecialWeight { get; set; }
-    public int SpecialCount { get; set; }
+    public double SpecialCount { get; set; }
     public decimal NaturalWeight { get; set; }
-    public int NaturalCount { get; set; }
+    public double NaturalCount { get; set; }
     public decimal NaturalWhiteWeight { get; set; }
-    public int NaturalWhiteCount { get; set; }
+    public double NaturalWhiteCount { get; set; }
     public decimal NaturalRedWeight { get; set; }
-    public int NaturalRedCount { get; set; }
+    public double NaturalRedCount { get; set; }
     public decimal ShortCutWeight { get; set; }
-    public int ShortCutCount { get; set; }
+    public double ShortCutCount { get; set; }
     public decimal ArtificialWeight { get; set; }
-    public int ArtificialCount { get; set; }
+    public double ArtificialCount { get; set; }
     public decimal ShortWeight { get; set; }
-    public int ShortCount { get; set; }
+    public double ShortCount { get; set; }
     public decimal LossWeight { get; set; }
     public decimal TotalWeight { get; set; }
     public decimal RemainingWeight { get; set; }
